+++ src/main/java/chess/daospring/GridDAO.java
@@ -1,22 +1,16 @@
-package chess.springdao;
+package chess.daospring;
 
-import chess.dao.ConnectionSetup;
 import chess.domain.grid.Grid;
 import chess.domain.grid.gridStrategy.NormalGridStrategy;
 import chess.domain.piece.Color;
 import chess.dto.GridDto;
-import chess.dto.response.ResponseCode;
-import chess.exception.ChessException;
 import org.springframework.jdbc.core.JdbcTemplate;
 import org.springframework.jdbc.support.GeneratedKeyHolder;
 import org.springframework.jdbc.support.KeyHolder;
 
-import java.sql.Connection;
 import java.sql.PreparedStatement;
-import java.sql.ResultSet;
 import java.sql.SQLException;
 import java.sql.Statement;
-import java.time.LocalDateTime;
 
 public class GridDAO {
     private JdbcTemplate jdbcTemplate;
